backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
numpy==2.4.4
pillow==12.2.0
openai==1.60.0
ultralytics==8.4.37
gradio==6.12.0
pydantic==2.10.1
python-multipart>=0.0.18
